# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fashion Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  user: User;
}

export interface UserUpdate {
  name?: string;
  avatar?: string;
  role?: string;
  active?: boolean;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: number;
  nameEn: string;
  nameAr: string;
  slug?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  parentCategoryId?: number | null;
  productCount?: number;
}

export interface CategoryInput {
  nameEn: string;
  nameAr: string;
  slug?: string;
  imageUrl?: string;
  /** @nullable */
  parentCategoryId?: number | null;
}

export interface CategoryUpdate {
  nameEn?: string;
  nameAr?: string;
  slug?: string;
  imageUrl?: string;
  /** @nullable */
  parentCategoryId?: number | null;
}

export interface ProductVariant {
  id: number;
  color: string;
  size: string;
  stockQuantity: number;
}

export interface ProductImage {
  id: number;
  imageUrl: string;
  isPrimary: boolean;
}

export interface Product {
  id: number;
  nameEn: string;
  nameAr: string;
  /** @nullable */
  descriptionEn?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  categoryId: number;
  categoryName?: string;
  vendorId: number;
  vendorName?: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  /** @nullable */
  sku?: string | null;
  featured?: boolean;
  active: boolean;
  averageRating?: number;
  reviewCount?: number;
  images?: ProductImage[];
  variants?: ProductVariant[];
  createdAt: string;
}

export interface VariantInput {
  color: string;
  size: string;
  stockQuantity: number;
}

export interface ProductInput {
  nameEn: string;
  nameAr: string;
  descriptionEn?: string;
  descriptionAr?: string;
  categoryId: number;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  sku?: string;
  featured?: boolean;
  active?: boolean;
  images?: string[];
  variants?: VariantInput[];
}

export interface ProductUpdate {
  nameEn?: string;
  nameAr?: string;
  descriptionEn?: string;
  descriptionAr?: string;
  categoryId?: number;
  price?: number;
  /** @nullable */
  salePrice?: number | null;
  sku?: string;
  featured?: boolean;
  active?: boolean;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productNameEn: string;
  productNameAr: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  size?: string | null;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  userId: number;
  userName?: string;
  totalPrice: number;
  paymentMethod: string;
  status: string;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  discount?: number | null;
  items?: OrderItem[];
  createdAt: string;
}

export interface OrderItemInput {
  productVariantId: number;
  quantity: number;
  price: number;
}

export interface OrderInput {
  paymentMethod: string;
  couponCode?: string;
  items: OrderItemInput[];
}

export interface OrderStatusUpdate {
  status: string;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface CartItem {
  variantId: number;
  productId: number;
  nameEn: string;
  nameAr: string;
  /** @nullable */
  imageUrl?: string | null;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  quantity: number;
  color: string;
  size: string;
  stockQuantity?: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  discount?: number;
  total: number;
}

export interface CartItemInput {
  variantId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface Review {
  id: number;
  productId: number;
  userId: number;
  userName?: string;
  /** @nullable */
  userAvatar?: string | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  rating: number;
  comment?: string;
}

export interface WishlistItem {
  productId: number;
  product: Product;
}

export interface Coupon {
  id: number;
  code: string;
  discountType: string;
  discountValue: number;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  usageLimit?: number | null;
  usageCount?: number;
  active: boolean;
}

export interface CouponInput {
  code: string;
  discountType: string;
  discountValue: number;
  startDate?: string;
  endDate?: string;
  usageLimit?: number;
}

export interface CouponUpdate {
  code?: string;
  discountType?: string;
  discountValue?: number;
  startDate?: string;
  endDate?: string;
  usageLimit?: number;
  active?: boolean;
}

export interface CouponValidateInput {
  code: string;
}

export interface Banner {
  id: number;
  titleEn: string;
  titleAr: string;
  /** @nullable */
  subtitleEn?: string | null;
  /** @nullable */
  subtitleAr?: string | null;
  imageUrl: string;
  /** @nullable */
  linkUrl?: string | null;
  active: boolean;
  sortOrder: number;
}

export interface BannerInput {
  titleEn: string;
  titleAr: string;
  subtitleEn?: string;
  subtitleAr?: string;
  imageUrl: string;
  linkUrl?: string;
  active?: boolean;
  sortOrder?: number;
}

export interface BannerUpdate {
  titleEn?: string;
  titleAr?: string;
  subtitleEn?: string;
  subtitleAr?: string;
  imageUrl?: string;
  linkUrl?: string;
  active?: boolean;
  sortOrder?: number;
}

export interface Notification {
  id: number;
  userId: number;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface AnalyticsSummary {
  totalRevenue: number;
  totalOrders: number;
  totalCustomers: number;
  totalProducts: number;
  revenueGrowth: number;
  ordersGrowth: number;
  pendingOrders?: number;
  lowStockCount?: number;
}

export interface SalesDataPoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface TopProduct {
  productId: number;
  nameEn: string;
  nameAr: string;
  /** @nullable */
  imageUrl?: string | null;
  totalSold: number;
  revenue: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface VendorSummary {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  lowStockProducts?: number;
}

export type ListUsersParams = {
role?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListProductsParams = {
search?: string;
categoryId?: number;
vendorId?: number;
minPrice?: number;
maxPrice?: number;
sizes?: string;
colors?: string;
featured?: boolean;
sortBy?: string;
page?: number;
limit?: number;
};

export type ListOrdersParams = {
status?: string;
userId?: number;
page?: number;
limit?: number;
};

export type GetSalesTimelineParams = {
period?: string;
};